pyproject.toml
[tool.poetry]
name = "guide-app"
version = "0.1.0"
description = ""
authors = ["nn"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.16.0"
django-cors-headers = "^4.7.0"
pillow = "^11.3.0"
django-filter = "^25.1"
requests = "^2.32.4"
djangorestframework-simplejwt = "^5.5.1"
requests-oauthlib = "^2.0.0"
django-admin-interface = "^0.30.1"
django-flat-theme = "^1.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
